+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `ListItem`: handle `aria-disabled=true` for an accessible disabled state
 -   `RadioButton`: handle `aria-disabled=true` for an accessible disabled state
 -   `Switch`: handle `aria-disabled=true` for an accessible disabled state
+-   `DisabledStateProvider` and `useDisabledStateContext` utilities to set disabled state in nested children
 
 ### Fixed